Adds helper for including in classes that are reacting
This allows the user to include `Riposte::Helper` in their classes and
either use a block or just get back the `Riposte::Response`.

```ruby
  #with block to call and pass the Response too
  react_to :success { |on| }

  #with no block
  react_to :success # #=> Riposte::Response.new(:success)
```

They will both technically return the response at this time, but that is
not always guaranteed to be true as of now.

Amos King @adkron <amos@binarynoggin.com>