Adds guard for auto running tests.
I don't really know how I feel about this being a dev dependency, but I
am opinionated so I'm putting it in anway. This also makes sure that I
can lock it down to a version for this project. Locking it down will
keep me from having to search for the right version of guard or update
the Guardfile as guard changes.

Amos King @adkron <amos@binarynoggin.com>